fix (improvement): Renaming the package to match to the final location
