+++ src/app/src/main/java/com/gabrielbmoro/programmingchallenge/ui/common/widgets/BubbleLoader.kt
@@ -1,4 +1,4 @@
-package com.gabrielbmoro.programmingchallenge.ui.components.compose
+package com.gabrielbmoro.programmingchallenge.ui.common.widgets
 
 import androidx.compose.animation.core.*
 import androidx.compose.runtime.Composable
